Updates tokenizer and index.html
Updates parse so that tag literals are not recognized until an end quote
is added. A token used to be classified as a tag literal as soon as any
character appeared after a quote. But this leads to premature filtering
of posts because of course nothing matches a half typed tag. Now the
tokenizer classifies unmatched quote strings as 'unknown' so they get
ignored by the parser/interpreter.

Updates index to include (commented out) test functions.
Also minor css changes.